Commit the workaround for Oktawave's still-unfixed power-on bug
There's still the bug that cloned VMs are booted automatically
and still worse, right after cloning the machine is reported as
powered off (it becomes powered on a few moments later), so we
cannot really determine whether the fix has been applied already.

Given it's over a month spent waiting for the fix, just publish
the workaround which ignores exceptions during the boot. Le sigh.